requirements.txt
Flask==2.1.2
flask-cors==3.0.10
Flask-RESTful==0.3.9
mysql-connector-python==8.0.29